feat: prospect pool API + Suggested tab, admin/dashboard/cache improvements
- Release 992 Salesforce-imported accounts (Mike's 988 + 4 buyer-owned) to prospect pool
- Add sf_account_id, import_priority, ownership_cooldown_until to Company model
- Build /api/prospects/pool endpoints (list, stats, detail, claim, dismiss)
- Add Suggested tab UI with card grid, search, priority filters, claim/dismiss actions
- Per-user cache keying in cached_endpoint decorator
- Fix cache TTL passthrough for sub-day values
- Admin and dashboard endpoint enhancements
- 26 new prospect pool tests, all passing

+++ app/cache/intel_cache.py
@@ -91,9 +91,9 @@ def get_cached(cache_key: str) -> dict | None:
     return None
 
 
-def set_cached(cache_key: str, data: dict, ttl_days: int = 7) -> None:
+def set_cached(cache_key: str, data: dict, ttl_days: float = 7) -> None:
     """Store data in cache with TTL."""
-    ttl_seconds = ttl_days * 86400
+    ttl_seconds = int(ttl_days * 86400)
 
     # Try Redis first
     r = _get_redis()
